Move TransportProtocol timeout extention:
 * Moved extension that depends on grpc
   from tests to protocol specific folder
   in lib

 * Added test-specific extension that pads
   timeout by 2s for use in tests
